+++ test/include/test.js
@@ -1,5 +1,4 @@
-
-describe('beforeEach-afterEach', () => {
+describe('include', () => {
 
     it('this test should work', () => {
 
@@ -7,4 +6,3 @@ describe('beforeEach-afterEach', () => {
     });
 
 });
-