+++ test/should-work/test.js
@@ -1,4 +1,3 @@
-
 describe('should work', () => {
 
     it('this test should work', () => {
@@ -7,4 +6,3 @@ describe('should work', () => {
     });
 
 });
-