+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Crapaud Changelog
 
+## [0.3.0] - 2020-12-24
+
+### Added
+ - Path can be a directory with a crapaud.json file instead.
+ - The `include` suite setting that will allow other suites to be included in a
+   run.
+
 ## [0.2.0] - 2020-12-23
 
 ### Added